Completion of getting point related data from the table
