+++ src/main/webapp/WEB-INF/web.xml
@@ -121,7 +121,7 @@
 
     <servlet>
         <servlet-name>PointSystem</servlet-name>
-        <servlet-class>com.cops.ntsf.controller.PointSystem</servlet-class>
+        <servlet-class>com.cops.ntsf.controller.PointSystemServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>PointSystem</servlet-name>
